# apps/image-rescale/src/main/java/com/newsnow/platform/imagerescale/adapters/driver/restapi/RescaleImageTaskGetController.java
package com.newsnow.platform.imagerescale.adapters.driver.restapi;

import com.newsnow.platform.imagerescale.core.ports.api.FindRescaleImageTask;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.UUID;
import java.util.logging.Level;
import java.util.logging.Logger;

@RestController
@RequestMapping("/task/{taskId}")
final class RescaleImageTaskGetController {

    private static final Logger LOGGER = Logger.getLogger(RescaleImageTaskGetController.class.getName());

    private final FindRescaleImageTask findRescaleImageTask;
    private final RescaleImageTaskRestApiMapper mapper;

    RescaleImageTaskGetController(FindRescaleImageTask findRescaleImageTask, RescaleImageTaskRestApiMapper mapper) {
        this.findRescaleImageTask = findRescaleImageTask;
        this.mapper = mapper;
    }

    @Tag(name = "Find Rescale Image Task")
    @Operation(
            summary = "Retrieve a rescale image task",
            description = "Endpoint to retrieve an existing rescale image task with its information saved in the system"
    )
    @ApiResponse(
            responseCode = "200",
            description = "Success loading the image task",
            content = @Content(schema = @Schema(
                    type = "object",
                    implementation = RescaleImageTaskRestApiDto.class
            ))
    )
    @ApiResponse(
            responseCode = "404",
            description = "Failed loading the image task",
            content = @Content(schema = @Schema(
                    type = "object",
                    implementation = RestApiErrorMessageResponse.class
            ))
    )
    @GetMapping(produces = "application/json")
    public ResponseEntity<RestApiResponse> perform(@PathVariable("taskId") UUID taskId) {
        var findRescaleImageTaskResult = findRescaleImageTask.apply(taskId);

        if (findRescaleImageTaskResult.hasErrors()) {
            findRescaleImageTaskResult.consumeErrors((error) ->
                    LOGGER.log(Level.WARNING, "[id: {0}] Error trying to find the rescale image task: {1}", new String[] {taskId.toString(), error})
            );
            return ResponseEntity.status(HttpStatus.NOT_FOUND)
                    .body(new RestApiErrorMessageResponse("Error trying to find the rescale image task, it may not have been found for taskId"));
        }

        RestApiResponse restApiResponse = mapper.toRestApiDto(findRescaleImageTaskResult.get());

        return ResponseEntity.status(HttpStatus.OK)
                .body(restApiResponse);
    }
}